APIs de pagamento/requirements.txt
Flask
python-dotenv
requests
stripe
mercadopago
